Исправлены тесты и некоторые баги

+++ ff-tvm/internal/service/repository.go
@@ -20,7 +20,7 @@ func (r *repositoryImpl) Create(ctx context.Context, service *Service) error {
 	).Scan(&service.ID)
 }
 
-func (r *repositoryImpl) GetByID(ctx context.Context, id int64) (*Service, error) {
+func (r *repositoryImpl) GetByID(ctx context.Context, id int) (*Service, error) {
 	service := &Service{}
 	err := r.db.QueryRowContext(ctx,
 		"SELECT id, name, public_key, private_key_hash FROM services WHERE id = $1",
@@ -32,7 +32,7 @@ func (r *repositoryImpl) GetByID(ctx context.Context, id int64) (*Service, error
 	return service, nil
 }
 
-func (r *repositoryImpl) GetPublicKey(ctx context.Context, id int64) (string, error) {
+func (r *repositoryImpl) GetPublicKey(ctx context.Context, id int) (string, error) {
 	var publicKey string
 	err := r.db.QueryRowContext(ctx,
 		"SELECT public_key FROM services WHERE id = $1",
@@ -44,7 +44,7 @@ func (r *repositoryImpl) GetPublicKey(ctx context.Context, id int64) (string, er
 	return publicKey, nil
 }
 
-func (r *repositoryImpl) GetPrivateKeyHash(ctx context.Context, id int64) (string, error) {
+func (r *repositoryImpl) GetPrivateKeyHash(ctx context.Context, id int) (string, error) {
 	var hash string
 	err := r.db.QueryRowContext(ctx,
 		"SELECT private_key_hash FROM services WHERE id = $1",
@@ -56,7 +56,7 @@ func (r *repositoryImpl) GetPrivateKeyHash(ctx context.Context, id int64) (strin
 	return hash, nil
 }
 
-func (r *repositoryImpl) GrantAccess(ctx context.Context, from, to int64) error {
+func (r *repositoryImpl) GrantAccess(ctx context.Context, from, to int) error {
 	_, err := r.db.ExecContext(ctx,
 		"INSERT INTO service_access (from_id, to_id) VALUES ($1, $2) ON CONFLICT DO NOTHING",
 		from, to,
@@ -64,7 +64,7 @@ func (r *repositoryImpl) GrantAccess(ctx context.Context, from, to int64) error 
 	return err
 }
 
-func (r *repositoryImpl) RevokeAccess(ctx context.Context, from, to int64) error {
+func (r *repositoryImpl) RevokeAccess(ctx context.Context, from, to int) error {
 	_, err := r.db.ExecContext(ctx,
 		"DELETE FROM service_access WHERE from_id = $1 AND to_id = $2",
 		from, to,
